[TASK] Extension kickoff
This change is the kickoff for the extension
having composer.json/ext_emconf.php in place
along with requiring specific version of the
deeplcom php client library and GitHub based
publish action as basic start.

With a later change automatic version rise
and releasing based on external package
can be done.